Changed error propagation
All connection related errors on the joystick will now make joystick do
some cleaning up and then terminate. Not currently attempting to get
error state of joystick thread back into main thread to send it to
remote as this is kind of unnecessary and a bit complicated. For now,
just send a command to restart joystick. Errors on launcher will now
cause the specific launch function to return to main which then looks
to receive the next command. Next job is to send the launcher errors
to remote. Can probably use select() for this on both sides