Removed a bunch of complexity. Added doc strings.
Launcher no longer has its own states as these are basically
duplicates of internal states of the class objects it handles. Now
launcher simply attempts to, for example, start the joystick and the
joystick checks its states itself and maybe returns an error. Much
cleaner. There is no longer Joystick object creation and deletion on
every start and stop joystick command. Instead, a Joystick is created
once and is reused. To do this I had to move TcpSocket creation from
the Joystick constructor to the connect function, so now only a
TcpSocket is recreated (the old one will be garbage collected) instead
of a whole Joystick. Escape from the update_values loop is no longer
done by closing the socket while it is blocked on recv, as this relies
on an exception. Now disconnect() simply sets the joystick state to
close_requested. recv exits on sending an additional pair of joystick
values. There is a new disconnect_internal function that does the
deallocation of resources, which was done previously using
disconnect. Loop break via the KILL command to the joystick has been
removed as STOP_JOYSTICK is sufficient for this. Unnecessary imports
just for error handling have been removed; TcpSocket, Joystick and
Launcher now only raise their own exceptions instead of propagating
other exceptions up. All resource handling in TcpSocket has been moved
to the close function so assignment to None no longer appears
everywhere. Error handling is looking good!