Bug fixes and added feature to unblock socket read and accept.
Fixed a bug where disconnecting the main command socket left the
joystick running. Now when closing the command socket, the launcher
function release_all() is called stops the joystick and clears any
other resources held. However when release_all() is called, joystick is very
likely to be stuck in a system call (recv) through
TcpSocket.read(). This is now solved by having a member socketpair in
TcpSocket. When read() is called, it blocks on select instead of
recv. A member unblock() method sends data from one end to the other
of the socketpair causing select to return. read then returns
None, just as it does for an EOF received. Joystick checks for None
and calls disconnect_internal. A bug was fixed where calling
START_JOYSTICK with an invalid port and then a valid port would not
start the joystick. To solve this, port number setting was moved from
the joystick contructor to the connect method. The external
disconnect() method no longer sets the state, it now only calls
TcpSocket unblock(). The CLOSE_REQUESTED state is now unused. Blocking
on select was also added to TcpSocket wait_for_connection() which
previously blocked on accept(). This is to enable solving in the
future, a bug where calling START_JOYSTICK and then STOP_JOYSTICK
immediately after causes the second command to hang as all commands
share the same thread and START_JOYSTICK causes blocking on
wait_for_connection(). The main function now also checks for a None
returned by read(). Sadly, two TcpSocket tests are currently failing
as I can't seem to figure out how to mock the select function properly.