Unit tests involving select fully working. Commands now use threads.
To prevent a command that leads to a block (like START_JOYSTICK which
leads to a blocked select() via wait_for_connection()) from not
allowing other commands (like STOP_JOYSTICK to terminate
wait_for_connection()) to get through, commands now are executed in
their own thread. These threads only live for the duration of the
command and have the benefit of being able to be shut down when blocked
in a syscall, using a different command. Was finally able to get the
mocking of select() to work properly. All current tests passing!