All current modules unit tested. Removed sock_lock in Joystick.
Apparently not closing the socketpair in TcpSocket is an uncleared
resource; close() now closes them. Error checking was surprisingly
missing in main(), added that in. Got rid of toggle_joystick from
launcher options as it is pretty pointless. Added an init test to
test_joystick. Wrote tests for main. Added some new tests in
test_tcpsocket, including one to check that unblock() works in a
concurrent setting. sock_lock was removed from joystick as there will
be concurrent access on TcpSockets to unblock them. Just need to make
sure that other methods in TcpSocket are not called concurrently. But
as of now, no known bugs.