Commit for pushing branch feature/interfaces to remote.
This branch will eventually be merged to dev/raspi. Currently have
migrated joystick (and arm) instruction receiving capability to a
'receiver' interface. Joystick (and arm) also inherit from an
'actuator' interface which will do all the work for sending
instructions to motors. Actuator has not yet been implemented. Current
approach to resource management is a resource_manager class. A
resource_manager will be taken as an argument to the constructor of a
Joystick (or RoboticArm) object. The resource_manager object will be
constructed by launcher so there will only be one resource_manager for
the duration of the program. The resource_manager will provide motor
access to joystick, arm, autonomous modes, etc. Unit tests have been
updated with the new approach. For now they are incomplete so don't
expect all of them to work on this commit.