Using START_JOYSTICK and STOP_JOYSTICK to do actuator stuff works.
Added a release(typename) method to resource manager for
dellallocation. However, this still depends on the user of the
resource to get rid of its own reference to the provided
resource. Changed motor_list in actuator from a list to a dict so that
the motors can be stored under their enum names as defined under
resource_manager. This is the bug fix for the previous commit where
the code was trying to index by enum name but motor_list was a simple
list. The release_motors method under Actuator has been given a quick
fix to ensure that things are working. The code below the return
statement there should be refactored (or maybe eliminated) in the next
commit. Acquiring wheel motors in Joystick is now done under the
start() method instead of __init__() so that Joystick function can be
restored with START_JOYSTICK after a STOP_JOYSTICK
command. ResourceManager's resources dict now indexes on the enum
value of a resource instead of a string. The resource_manager tests
have been ammended to match this. A bug that exists now is, when
Joystick is closed by interrupting the connection instead of calling
STOP_JOYSTICK, the actuator update_motors thread keeps running as it
has no way of knowing that receiver functions have stopped.