Fixed bug where interrupting the connection left actuator active.
Fixed this by introducing the receiver abstract method
run_on_connection_interrupted which runs when a TcpSocketError occurs
on read or a RST is received (empty data). This is implemented by
classes deriving from receiver like Joystick or RoboticArm, so that
they can run some optional cleanup. In this case, the cleanup is
releasing the motors. Currently implemented on Joystick. Will
implement on RoboticArm when writing the appropriate tests in the future.