Actuator now waits for a new value without continuously updating.
This is done using Threading.Condition. Actuator has a member
condition which it calls wait() on, which blocks the thread until
notify() is called on the condition from elsewhere. Joystick calls
notify() after receiver updates its values which unblocks
update_motors so that actuator can proceed to write the new values to
motors. release_motors also calls notify() on the condition. On
unblocking, update_motors checks if release_motors was called, and if
it was, calls release_motors on any remaining motors and exits the
loop. This is because the current implementation assumes that an
actuator cannot carry out its function without a full set of
motors. This is subject to change later. Joystick no longer has a
value_lock as the condition provides locking as well.