Combined Motor Code
Combined the motor code for robotic arm and wheel motors into one .py, only
included the code for one motor in the document as unsure how data is fed
into the file. If hardware PWM needed for motors, can copy package motor.py
in /usr/local/lib/python3.5/dist-packages/ and use that. TODO: Unit testing
on Servos and on the motors. Need to create a stream of test data in order
to test the motors.