Wrote tests for configuration. More debugging info is now printed.
Configuration has a few unit tests for it now, but is not really
complete yet. In order to exit cleanly on connection lost, a few
changes are made. First, all controllers now have an is_running which
which is checked by launcher before it tries to kill them in
release_all, preventing multiple stop commands executing. The
"running" state is active as soon as the controller starts waiting for
a connection (if it inherits from receiver) and also when the start
command of the controller is executed, and is made inactive on fully
stopping the controller. For the inheritance from receiver to work,
the controller now has to implement the set_state_as_running abstract
method of receiver (as it is needed by receiver to change the state to
"running" on connection initiation). This is a bit clunky, so it might
be changed later but things seem to work. This means all controllers
now need their own enum class to define states. Clunky!! As a degree
of redundancy, the stop commands on Joystick and RoboticArm only
execute if the state is "running". Will probably get rid of this later
to expose any hidden bugs with the mechanism. More error handling is
done with motors_interface now. Still have only tested error messages
on Laptop mode. For now at least, errors in configuration related to
motors propagate through the motors_interface and then through
resource_manager. To prevent errors at the import statement for
resource_manager in start_rover, all configuration activities in
resource_manager are moved to an initialise() function, which is
explicitly called by start_rover and error handled. Joystick and arm
tests were updated (although they now need a thorough go-through as
quite a few things have been added since the tests were written) to
make them pass with the new state information added.