Better connection termination handling, with better error messages.
Launcher now checks if a controller is running before calling their
respective kill functions. It no longer passes stop commands through
to already stopped controllers and prints a message if this is
attempted. To do this, receiver now has a new connection_active()
function that returns true if the connection state is not
"CLOSED". This is called by Joystick and RoboticArm in their close()
function to see if the connection requires termination before
attempting it.