Added camera, CameraUser, and a full camera stream mode.
The camera resource is pretty much complete. It uses gstreamer via the
opencv API for it, to get frames. The camera can be made to start
capturing frames and storing them, which are retrieved via a getter
for use. The camera settings such as framerate, width and height are
set through the settings file. The configure file has been updated
with a CameraConfiguration, similar to MotorConfiguration to give
access to camera-specific settings. Anything that requires access to
the camera will inherit from the CameraUser interface. This provides
methods to gain access to the camera via resource_manager, and also
provides streaming capability, again through gstreamer. The IP
address, port, stream framerate, stream width and height are all set
in the settings file. A simple camera feed has now been implemented by
inheriting from this. It does no processing of the frames (simply
streams it). This is available under
camera_feed/stream_default.py. Tests have been written for both the
CameraUser interface and camera resource. Launcher, parser and
start_rover have been updated with the new commands for starting and
stopping the simple stream. One issue with this implementation is that
storing frames by the camera resource and retrieving them again by the
CameraUser interface is slow and causes close to a second of lag on
the stream, even without mutexes (relying on the GIL alone). This will
need to be fixed, perhaps by bypassing storing of frames
somehow. Perhaps can explore writing the resource and interface as C extensions.