Camera no longer runs continuously in thread.
Reduced overhead by only capturing a frame when needed. This removes
the need to run the camera in its own thread and also removes the need
for concurrency locks on frames. Resource manager now takes care of
initialising the camera. cv2's cap.release() is called by resource
manager when the camera use count decreases to zero, VideoCapture is
called when the camera is acquired for the first time. The program now
checks for Mac computers (in LAPTOP mode) via sys.platform, removing
the need for an additional mode.