Fixes for choppy video, switched to raspivid, frames in RGB.
Largest frame size that works with streaming on the Raspberry Pi was
found to be 360x200. This is now the default setting under CAMERA in
the settings file, along with 60fps framerate for both capture and
stream. Currently using raspivid as the video source in RASPBERRYPI
mode, although v4l2src works equally well. Removed h264parse from the
streaming pipeline which previously caused greyed out frames in the
stream. Previously frames were streamed in BGR. Switched to RGB.