Settings file can now be specified at commandline.
Configuration is (probably) thread safe as now it always acquires a
lock when parsing the settings files. Motor configuration is also now
global. All future configuration managers should be global so that the
settings file can be changed concurrently via locking. pwm_hardware
settings have been removed as they are unused. Resource manager checks
with the Configuration managers if they have been initialised. If they
have not, an error is thrown and the application terminates. The
initialisation is done in main() and the resource manager is made to
check afterwards, so this error will only occur if somehow resource
manager is initialised before the configuration manager is. A
laptop_settings xml file has been included as an alternative.