Added separate debug settings file for raspberry pi.
The laptop mode settings file needs lots of changes to use on the
raspberry pi for debugging purposes, so might as well have a separate
debug xml file for the raspberry pi. For this, there is a new 'mode'
called RASPBERRYPI_NO_MOTORS, meaning the mock motors are used instead
of importing the real motor libraries, so that the program can run
without the motors connected.