Added Resource base class.
This lets resources be specified without requiring any framework code
changes. Works very much like OpModes where the files/folders where
the resource classes may be found are listed in the settings file. To
use this, resource manager has been changed fairly significantly. The
resource class has a policy member that specifies whether the resource
should be acquired uniquely by a given mode, or if sharing between
active modes is allowed and causes no conflict. The motors interface
doesn't work yet in this configuration to differentiate mock motors
from the real libraries. Something not implemented yet but has to be
in the future is automatically invalidating names associated with a
resource object, after the using mode has called release() on that
resource so that any attempt to use that name afterwards causes an
immediate error. At the moment, resource management is entirely
dependent on the modes always doing the right thing.