Got rid of motors_interface.
Simulating motors is now done by having the mock motors use the
registered name of the real motors and specifying it in the settings
file instead of the actual motors. This works because opmodes look for
resources through their registered names. All tests are
passing. Probably can completely get rid of the raspberrypi/laptop
mode, etc and completely rely on settings files to specify things.