Bare-minimum to get camera working with new opmodes.
Tests are still incomplete but the default stream mode is rewritten as
an opmode. The need for hardware mode settings has been removed by
specifying camera device in settings (v4l2src, rpicamsrc,
etc). CameraConfiguration is updated to accommodate this. Camera was
rewritten as a Resource with shared_init and shared_deinit functions
to start and stop capture.