Added diagnostics - all print statements optionally go to a socket.
If enabled in settings, using dg.print instead of print will print
any messages to console as well as send them out on the specified tcp
port. Diagnostics has an internal buffer of a fixed length that holds
a certain number of messages if the connection is not yet established,
and prints them out when the connection is established. No tests
written for this yet.