Basic (crude) working remote resources. Added client_socket class.
Remote procedure calls for accessing resources on an attached unit
works on a basic level. Code is quite messy at the moment - needs
refining and removing random global states and unused stuff. RPC was
tested with remote wheel motors and local joystick mode. Lots of
things are incomplete such as releasing threads and resource
management is unchecked at the moment. Likely next commit will fix
most of this stuff. offload.py has been renamed unit.py to reflect
that it's not purely for offloading (also offloading is unimplemented
at the moment). Added laptop_settings2.xml to be used with running the
program with --as-unit to complement the main unit running with
laptop_settings.xml, for testing purposes. Obviously, unit tests are
incomplete/haven't been written. Added a client_socket class which is
essentially a client version of tcpsocket. It is currently
unused. Initially planned to use it with the RPC but RPyC has its own
socket mechanism so that is being used instead.