WIP: unresolved bug involving remote resource registration.
Configuration stores unitname. Resource manager now differentiates
between local and remote resources - so that units on which a given
resource is declared can be used locally on that unit. Lots of
debugging print statements. There is an unresolved bug where a remote
resource registers on with the local resource manager and somehow
tries to initialise as a local resource on the main unit. start_rover
no longer loads opmodes on units other than the main unit.