Fixed remote resource bug. Added unit disconnect cleanup code.
Remote resource bug turned out to be that resource.policy is a netref object because it was assigned to by the main unit (while resource was initialised in an attached unit) which meant that it was trying to communicate to the main unit despite being a local resource. Removed a bunch of debug print statements.

Added stuff under on_disconnect of the main server. This removes any remote resources of the disconnecting unit, from the remote resource list (in Resources) and from the resource manager. This makes it possible to have the unit disconnect and reconnect multiple times. However currently disconnection while the resource is acquired is unhandled. This requires resource manager to know the opmode using the resource and to stop the mode. Likely next commit will have this dealt with.