Implemented Goals - starting and stopping working perfectly.
goal_samples.py contains goal for picking up samples. At the moment
the camera is concurrently accessed which means a mutex needs to be
used at the camera level - this is unbearably latent and an
alternative must be found. Added cascade files. The Samples goal will
make use of a 'CVEngine' which provides the object recognition capabilities.